perf(messaging): decouple from code_puppy.tools to unblock cold-start chokepoint
Every plugin that imports from code_puppy.messaging currently drags in
~870 modules (pydantic_ai, playwright, prompt_toolkit, rich, httpx, ...)
because messaging.rich_renderer imports two symbols from code_puppy.tools
at module scope, which forces code_puppy/tools/__init__.py to execute the
full static tool registry (browser->playwright, image_tools->PIL,
agent_tools->rapidfuzz, and dozens more).

This change breaks that coupling without touching the tools registry:

1. Move subagent_context out of code_puppy.tools/ to code_puppy/, since
   it's pure contextvars bookkeeping with zero tool dependencies.
2. Leave a re-export shim at the old path so external plugins and
   monkeypatch-based tests keep working.
3. Update the two module-scope import triggers in messaging/rich_renderer
   and messaging/spinner to point at the new canonical location.
4. Defer the one remaining tools.common symbol (format_diff_with_colors)
   to the _render_diff call site — it's only used inside a single
   post-tool-call render path.

Result (measured across all 51 builtin plugins):

  Total modules loaded during plugin registration: 26,898 -> 16,646
  (-10,252 modules, -38%)

  Nineteen plugins each drop ~622 modules (theme, aws_bedrock,
  customizable_commands, etc.), and every plugin that touches messaging
  benefits without any change to plugin code.

The tools registry itself is unchanged — this is a decoupling, not a
lazy-import refactor of the registry. Follow-up work can lazy-load
individual tools (option A in the scoping report at
~/projects/startup-perf-scoping-2026-07-27.md).

Test bookkeeping:

* tests/test_rich_renderer.py previously stubbed code_puppy.tools.common
  via sys.modules.setdefault. That stubbing only worked by accident on
  main because rich_renderer's own eager import of tools.common ran
  first and populated the real module. With the eager import gone, the
  incomplete stub would shadow the real module. Removed the stubs and
  documented why — the real module is cheap to load for one test file.

* tests/messaging/spinner/test_spinner_shim.py points its monkeypatch
  at the new canonical is_subagent location (spinner does a
  function-scope from-import each call, so it picks up the mock).

2,016 tests pass across messaging/, agents/, tools/, and plugins/.
